Add label for proxy
Some deployments may have multiple twemproxy processes on the same host, all of which are proxying the same clusters and servers. Similar to https://github.com/xavierholt/twemproxy_exporter/pull/4 we then run into issues, where a single counter may be shared across multiple proxies and count incorrectly. The label `twemproxy` is not sufficient in this case, because it comes from twemproxy's stats and is set to the current hostname (not configurable).